\name{adalasso-package}
\alias{adalasso-package}
\alias{adalasso}
\docType{package}
\title{
Use the adaptive lasso for shrinkage and selection in a generalized linear model
}
\description{
Iterate between glmnet (for shrinkage/selection) and glm (for adaptive penalty weights) to get an adaptive lasso estimate of coefficients in a GLM where p>n.
}
\details{
\tabular{ll}{
Package: \tab adalasso\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2012-07-03\cr
License: \tab GPL-3\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Wesley Brooks

Maintainer: Wesley Brooks <wrbrooks@usgs.gov>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
